ui: remove leftnav, put controls in main titlebar
