+++ ui/layout.jsx
@@ -1,19 +1,31 @@
 'use babel'
 import React from 'react'
+import { Link } from 'react-router'
 import app from './lib/app'
-import LeftNav from './views/leftnav'
 import Notifications from './com/msg-list/notifications'
 import Bookmarks from './com/msg-list/bookmarks'
 import ModalFlow from './com/modals/flow'
 import ProfileSetup from './com/forms/profile-setup'
 import FollowNearby from './com/forms/follow-nearby'
 import PubInvite from './com/forms/pub-invite'
+import Issues from './com/issues'
 
 const SETUP_LABELS = [<i className="fa fa-user"/>, <i className="fa fa-wifi"/>, <i className="fa fa-cloud"/>]
 const SETUP_FORMS = [ProfileSetup, FollowNearby, PubInvite]
 const RIGHT_NAVS = {
   notifications: Notifications,
   bookmarks: Bookmarks
+}
+
+class NavLink extends React.Component {
+  render() {
+    const selected = (this.props.to === this.props.location)
+    const cls = 'ctrl '+(selected?'selected':'')
+    const count = this.props.count ? <div className="count">{this.props.count}</div> : ''
+    if (!this.props.children)
+      return <Link className={cls} to={this.props.to}><i className={'fa fa-'+this.props.icon} /><span className="label">{this.props.label}</span> {count}</Link>
+    return <Link className={cls} to={this.props.to}>{this.props.children}</Link>
+  }
 }
 
 export default class Layout extends React.Component {
@@ -35,6 +47,7 @@ export default class Layout extends React.Component {
     return {
       rightColumn: (this.state) ? this.state.rightColumn : false,
       isWifiMode: app.isWifiMode,
+      indexCounts: app.indexCounts||{},
       user: app.user,
       users: app.users,
       setupIsOpen: app.user.needsSetup,
@@ -53,11 +66,11 @@ export default class Layout extends React.Component {
   onClickBack() {
     window.history.back()
   }
-  
+
   render() {
-    const onToggleRightNav = (id) => () => {
-      this.toggleRightNav(id)
-    }
+    const location = this.props.location.pathname
+    const isWifiMode = this.state.isWifiMode
+    const onToggleRightNav = (id) => () => { this.toggleRightNav(id) }
     const composing = this.state.isComposerOpen
     const RightNavView = (this.state.rightColumn) ? RIGHT_NAVS[this.state.rightColumn] : null
     return <div className="layout-rows">
@@ -65,6 +78,13 @@ export default class Layout extends React.Component {
       <div className="toolbar flex">
         <div className="flex-fill">
           <a className="ctrl back" onClick={this.onClickBack}><i className="fa fa-angle-left" /></a>
+          <div className="nav">
+            <NavLink to="/" location={location} icon="newspaper-o" label="Feed" />
+            <NavLink to="/inbox" location={location} icon="inbox" label="Inbox" count={this.state.indexCounts.inboxUnread} />
+            <NavLink to="/profile" location={location} icon="users" label="Contacts" />
+            <NavLink to="/sync" location={location} icon={isWifiMode?'wifi':'globe'} label='Network' />
+            <Issues />
+          </div>
         </div>
         <div>
           <div className="search"><i className="fa fa-search" /><input /></div>
@@ -73,14 +93,6 @@ export default class Layout extends React.Component {
         </div>
       </div>
       <div className="layout-columns">
-        <LeftNav
-          location={this.props.location.pathname}
-          isWifiMode={this.state.isWifiMode}
-          userid={this.state.user.id}
-          names={this.state.users.names}
-          friends={this.state.user.friends}
-          following={this.state.user.nonfriendFolloweds}
-          followers={this.state.user.nonfriendFollowers} />
         <div id="mainview">{this.props.children}</div>
         { (RightNavView) ? <div id="rightnav"><RightNavView /></div> : '' }
       </div>
