+++ ui/less/com/msg-view/summary.less
@@ -2,7 +2,6 @@
   display: flex;
   padding: 10px;
   overflow: hidden;
-  margin-right: 50px;
 
   .ctrls {
     flex: 0 0 24px;
