+++ ui/less/com/toolbar.less
@@ -39,18 +39,49 @@
     }
   }
   .ctrl {
+    position: relative;
     display: inline-block;
     padding: 0 12px;
     font-size: 21px;
     text-align: center;
     color: rgb(104, 120, 144);
-    &:hover {
+    &:hover, &.selected {
       color: rgb(57, 65, 77);
+    }
+
+    .count {
+      position: absolute;
+      right: 0px;
+      top: -2px;
+      background: #687890;
+      color: #FFFFFF;
+      width: 16px;
+      height: 16px;
+      border-radius: 10px;
+      font-size: 12px;
     }
   }
   .back {
     width: 50px;
     padding: 0;
+  }
+  .nav {
+    display: inline-block;
+    .ctrl i {
+      display: none;
+      padding: 0 8px;
+    }
+    .ctrl .label {
+      padding: 0 4px;
+    }
+    @media (max-width: 960px) {
+      .ctrl i {
+        display: inline;
+      }
+      .ctrl .label {
+        display: none;
+      }
+    }
   }
 
   .search {
