+++ ui/less/main.less
@@ -45,12 +45,9 @@
 @import "transitions/slideup.less";
 @import "transitions/fade.less";
 
-@import "views/leftnav.less";
 @import "views/rightnav.less";
-@import "views/composer-sidepanel.less";
 @import "views/newsfeed.less";
 @import "views/inbox.less";
-@import "views/bookmarks.less";
 @import "views/data.less";
 @import "views/msg.less";
 @import "views/sync.less";
