+++ ui/less/views/rightnav.less
@@ -1,6 +1,8 @@
 #rightnav {
+  width: 300px;
+  border-left: 1px solid #ccc;
+  background: #fff;
   .notifications {
-    background: #fff;
     .toolbar {
       .tabs {
         display: flex;
@@ -11,8 +13,7 @@
       }
     }
     .msg-view.notification {
-      width: 275px;
-      border-left: 1px solid #ccc;
+      width: 290px;
     }
   }
 }
